Split estimates in two.
