+++ resource_estimator/src/estimates/data/job_params.rs
@@ -7,9 +7,10 @@ use serde::{Deserialize, Serialize};
 
 use super::super::{
     error::InvalidInput::{self, InvalidErrorBudget},
-    modeling::{ErrorBudget, PhysicalQubit, ProtocolSpecification},
+    modeling::{PhysicalQubit, ProtocolSpecification},
     LogicalResourceCounts,
 };
+use crate::estimates2::ErrorBudget;
 
 use super::{tfactory::TFactoryDistillationUnitSpecifications, Constraints};
