+++ resource_estimator/src/estimates/modeling.rs
@@ -1,11 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-mod error_budget;
 mod fault_tolerance;
-mod logical_qubit;
 mod physical_qubit;
-pub use error_budget::*;
 pub use fault_tolerance::*;
-pub use logical_qubit::*;
 pub use physical_qubit::*;
