+++ resource_estimator/src/estimates/modeling/fault_tolerance.rs
@@ -7,7 +7,7 @@ mod tests;
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 
-use crate::estimates::stages::physical_estimation::ErrorCorrection;
+use crate::estimates2::ErrorCorrection;
 
 use super::{
     super::{
