+++ resource_estimator/src/estimates/optimization.rs
@@ -3,8 +3,6 @@
 
 mod code_distance_iterators;
 mod distillation_units_map;
-mod population;
 mod tfactory_exhaustive;
 
-pub(crate) use population::{Point2D, Population};
 pub(crate) use tfactory_exhaustive::TFactoryBuilder;
