+++ resource_estimator/src/estimates/optimization/tfactory_exhaustive/tests.rs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-use crate::estimates::stages::physical_estimation::{ErrorCorrection, Factory};
+use crate::estimates2::{optimization::Point2D, ErrorCorrection, Factory};
 
 use super::{
     super::super::{
@@ -10,7 +10,6 @@ use super::{
         stages::tfactory::{TFactory, TFactoryDistillationUnitTemplate},
     },
     find_nondominated_population, find_nondominated_tfactories,
-    population::Point2D,
 };
 use std::{rc::Rc, time::Instant};
