+++ resource_estimator/src/estimates/stages.rs
@@ -2,5 +2,4 @@
 // Licensed under the MIT License.
 
 pub mod layout;
-pub mod physical_estimation;
 pub mod tfactory;
