+++ resource_estimator/src/estimates2.rs
@@ -1,0 +1,12 @@
+mod error_budget;
+pub use error_budget::ErrorBudget;
+mod physical_estimation;
+pub use physical_estimation::{
+    ErrorCorrection, Factory, FactoryBuilder, PhysicalResourceEstimation,
+    PhysicalResourceEstimationResult,
+};
+mod layout;
+mod logical_qubit;
+pub use layout::Overhead;
+pub use logical_qubit::LogicalQubit;
+pub mod optimization;
