+++ resource_estimator/src/estimates2/optimization.rs
@@ -1,0 +1,2 @@
+mod population;
+pub use population::{Point, Point2D, Point4D, Population};
