+++ resource_estimator/src/estimates2/optimization/population.rs
